requirements.txt
streamlit==1.28.1
openai-whisper==20231106
ffmpeg-python==0.2.0
torch==2.1.0+cu118